fix: rename folder with context
